Add support for entering years and bug fixes. Closes #4 and fixes #7
In order to make entering times without a colon easier (1130), years
are assumed to be in the 21st century, so only years beginning with "20"
will be recognized. If you must enter some other year, you can use
the format "mm/dd/yyyy" for more leniancy in years.

This means that dec 5 2100 will parse as Dec 5 at 9pm, while
dec 5 2030 will parse as Dec 5, 2030.